Move useFormInput hook to its own hooks file and import
